% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{barplotMSSignal}
\alias{barplotMSSignal}
\title{Barplot of ms-signal per column}
\usage{
barplotMSSignal(
  eset,
  col = as.character(.getConditionColors(eset)[pData(eset)$condition, ]),
  method = c("sum", "sharedSignal"),
  cex.lab = 1.25,
  cex.axis = 1.25,
  cex.names = 0.9,
  labels = rownames(pData(eset)),
  ...
)
}
\arguments{
\item{eset}{expressionSet}

\item{col}{default condition colors}

\item{method}{c("median","sum","sharedSignal")}

\item{cex.lab}{default 1.25}

\item{cex.axis}{default 1.25}

\item{cex.names}{default 0.9}

\item{labels}{labels}

\item{...}{see plot}
}
\description{
Barplot of ms-signal per column
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
