% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{cvBoxplot}
\alias{cvBoxplot}
\title{C.V. boxplot}
\usage{
cvBoxplot(
  eset,
  col = as.character(.getConditionColors(eset)[unique(pData(eset)$condition), ]),
  cex.names = 0.9,
  cex.axis = 1.25,
  cex.lab = 1.25,
  ylab = "C.V. (\%)",
  ...
)
}
\arguments{
\item{eset}{ExpressionSet}

\item{col}{col}

\item{cex.names}{default 0.9}

\item{cex.axis}{default 1.25}

\item{cex.lab}{default 1.25}

\item{ylab}{C.V.}

\item{...}{see plot}
}
\description{
C.V. boxplot
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
