% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserOptions.R
\name{expDesignTagToExpDesign_file_2}
\alias{expDesignTagToExpDesign_file_2}
\title{Create experimental design data.frame from user input string}
\usage{
expDesignTagToExpDesign_file_2(t, ...)
}
\arguments{
\item{tag}{tag}

\item{expDesignDefault}{data.frame}
}
\value{
data.frame describing experimental design
}
\description{
Create experimental design data.frame from user input string
}
\details{
Example Files for the different file formats
MaxQuant File:
Name	Experiment	Groups	Type
LFQ intensity 6892 - Calcium -1	6892 - Calcium -1	6892 - Calcium	C
LFQ intensity 6892 + Calcium -1	6892 + Calcium -1	6892 + Calcium	T
LFQ intensity 7648 - Calcium -1	7648 - Calcium -1	7648 - Calcium	T
LFQ intensity 7648 - Calcium -2	7648 - Calcium -2	7648 - Calcium	T
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
