% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getAllCV}
\alias{getAllCV}
\title{Calculate Coefficiant of Variance per feature (Relative standard Deviation) per Condition}
\usage{
getAllCV(eset, l2 = TRUE)
}
\arguments{
\item{eset}{ExpressionSet}
}
\value{
data.frame of CVs per condition
}
\description{
Calculate Coefficiant of Variance per feature (Relative standard Deviation) per Condition
}
\details{
CV = sd / mean
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
\code{\link{getCV}}
}
