% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getAllEBayes}
\alias{getAllEBayes}
\title{Perform statistical test (mderated t-test), comparing all case to control}
\usage{
getAllEBayes(
  eset = eset,
  adjust = F,
  log = T,
  method = "pairwise",
  adjustFilter = matrix(F, nrow = nrow(eset), ncol =
    length(levels(pData(eset)$condition)) - 1)
)
}
\arguments{
\item{eset}{ExpressionSet}

\item{adjust}{TRUE/FALSE adjust for multiple testing using Benjamini & Hochberg  (1995) method}

\item{log}{T/F log-transform expression values}

\item{method}{c("all","pairwise")}

\item{adjustFilter}{matrix T/F do not adjust for multiple testing}
}
\value{
data.frame of pvalues per condition comparison
}
\description{
Perform statistical test (mderated t-test), comparing all case to control
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
Empirical Bayes method, Smyth (2004), \url{http://www.ncbi.nlm.nih.gov/pubmed/16646809}
}
\seealso{
\code{\link[limma]{eBayes}}
}
