% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getIBAQEset}
\alias{getIBAQEset}
\title{Calculate intensity-based absolute-protein-quantification (iBAQ) metric per protein}
\usage{
getIBAQEset(
  eset,
  proteinDB = NA,
  peptideLength = c(5, 36),
  nbMiscleavages = 0,
  proteaseRegExp = .getProteaseRegExp("trypsin")
)
}
\arguments{
\item{eset}{protein level ExpressionSet}

\item{proteinDB}{list protein sequneces}

\item{peptideLength}{peptide length interval (to get number of peptides used for normalization)}

\item{nbMiscleavages}{number of mis-cleavages allowed when digesting protein sequneces in silico (to get number of peptides used for normalization)}

\item{proteaseRegExp}{protease Reg Exp cleavage rule}
}
\value{
ExpressionSet
}
\description{
Calculate intensity-based absolute-protein-quantification (iBAQ) metric per protein
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
Global quantification of mammalian gene expression control, Schwanhausser (2011), \url{http://www.ncbi.nlm.nih.gov/pubmed/21593866},
Critical assessment of proteome-wide label-free absolute abundance estimation strategies. Ahrne (2013), \url{http://www.ncbi.nlm.nih.gov/pubmed/23794183}
}
