% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getModifProteinCoordinates}
\alias{getModifProteinCoordinates}
\title{Get modification coordinates on protein}
\usage{
getModifProteinCoordinates(modifAnnot, peptideSeq, proteinSeq, format = 1)
}
\arguments{
\item{modifAnnot}{modifcation as annotated by progenesis. E.g. '\link{15} Phospho (ST)|\link{30} Phospho (ST)'}

\item{peptideSeq}{peptide sequence}

\item{proteinSeq}{protein sequence}

\item{format}{c(1,2) 1. progenesis 2. scaffold}
}
\value{
vector of protein coordinates (mmodification residue number)
}
\description{
Get modification coordinates on protein
}
\details{
NA
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
