% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getPeptides}
\alias{getPeptides}
\title{Digest protein}
\usage{
getPeptides(
  proteinSeq,
  proteaseRegExp = .getProteaseRegExp("trypsin"),
  nbMiscleavages = 0
)
}
\arguments{
\item{proteinSeq}{protein sequence}

\item{proteaseRegExp}{protease Regular Expression}

\item{nbMiscleavages}{default 0}
}
\value{
vector of peptides
}
\description{
Digest protein
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
