% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getScoreCutOff}
\alias{getScoreCutOff}
\title{Get score cutoff for a given fdr cut-off}
\usage{
getScoreCutOff(scores, isDecoy, fdrCutOff = 0.01)
}
\arguments{
\item{scores}{peptide/protein identficationscore}

\item{isDecoy}{vector of TRUE/FALSE}

\item{fdrCutOff}{\link{0,1}}
}
\value{
scoreCutoff
}
\description{
Get score cutoff for a given fdr cut-off
}
\details{
NA
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
