% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{isStrippedACs}
\alias{isStrippedACs}
\title{Check if ACs are in "non-stripped" uniprot format e.g. "sp|Q8CHJ2|AQP12_MOUSE"}
\usage{
isStrippedACs(acs)
}
\arguments{
\item{acs}{accession numbers}
}
\value{
boolean TRUE/FALSE
}
\description{
Check if ACs are in "non-stripped" uniprot format e.g. "sp|Q8CHJ2|AQP12_MOUSE"
}
\details{
TRUE if less than 10\% of ACs contain a "|" character
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
