% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kinaseMotif}
\alias{kinaseMotif}
\title{Kinase motifs}
\format{
A data frame with 175 rows and 2 variables:
\describe{
\item{motif}{kinase motif)}
\item{kinase}{kinase}
}
}
\usage{
kinaseMotif
}
\description{
Human Protein Reference Database Serine/Threonine motifs
http://www.hprd.org/serine_motifs
The variables are as follows:
}
\keyword{datasets}
