% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{maPlotSQ}
\alias{maPlotSQ}
\title{ma-plot}
\usage{
maPlotSQ(
  eset,
  sample = colnames(exprs(eset))[1],
  cex.lab = 1.5,
  cex.axis = 1.5,
  lwd = 2,
  pch = 1,
  col = rgb(0, 100, 0, 50, maxColorValue = 255),
  ...
)
}
\arguments{
\item{eset}{ExpressionSet}

\item{sample}{selected condition}

\item{cex.lab}{default 1.5}

\item{cex.axis}{default 1.5}

\item{lwd}{default 2}

\item{pch}{default 1}

\item{col}{green transparent}

\item{...}{see plot}
}
\description{
ma-plot
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
