% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserOptions.R
\docType{data}
\name{option_list}
\alias{option_list}
\title{Command Line Option List}
\format{
An object of class \code{list} of length 36.
}
\usage{
option_list
}
\description{
Command Line Option List
}
\keyword{datasets}
