% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{pairsAnnot}
\alias{pairsAnnot}
\title{Plot lower triangle Pearsons R^2. Diagonal text, upper triangle all against all scatter plots with lm abline}
\usage{
pairsAnnot(
  data,
  textCol = rep(1, ncol(data)),
  diagText = colnames(data),
  col = rgb(0, 100, 0, 50, maxColorValue = 255),
  isHeatCol = F,
  cexTxt = 2,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{textCol}{text color}

\item{diagText}{diagnoal text}

\item{col}{dot col}

\item{isHeatCol}{heat colors}

\item{cexTxt}{cex txt}

\item{...}{see plot}
}
\description{
Plot lower triangle Pearsons R^2. Diagonal text, upper triangle all against all scatter plots with lm abline
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
