% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser.R
\name{parseDIANNPeptideMeasurementCsv}
\alias{parseDIANNPeptideMeasurementCsv}
\title{Parse DIANN Peptide Measurement after using the DIANN_converter}
\usage{
parseDIANNPeptideMeasurementCsv(
  file,
  expDesign = expDesign,
  method = "auc",
  expressionColIndices = .getProgenesisCsvExpressionColIndices(file, method = method),
  exclusivePeptides = F
)
}
\arguments{
\item{file}{path to DIANN Peptide Conerter csv file}

\item{expDesign}{experimental design data.frame}

\item{method}{auc (area under curve) or spc (spectral count)}

\item{expressionColIndices}{default .getProgenesisCsvExpressionColIndices()}

\item{specificPeptides}{T/F keep specific peptides only}
}
\value{
ExpressionSet object
}
\description{
Parse DIANN Peptide Measurement after using the DIANN_converter
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
\code{\link[Biobase]{ExpressionSet}}
}
