% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser.R
\name{parseDiaNNProteinGroupTxt}
\alias{parseDiaNNProteinGroupTxt}
\title{Parse DIA-NN files Protein Group Converter tsv}
\usage{
parseDiaNNProteinGroupTxt(file = file, expDesign = expDesign, method = "auc")
}
\arguments{
\item{file}{path to DIA-NN Converter tsv file}

\item{expDesign}{experimental design data.frame}

\item{method}{auc (area under curve) or spc (spectral count)}
}
\value{
ExpressionSet object
}
\description{
Parse DIA-NN files Protein Group Converter tsv
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
\code{\link[Biobase]{ExpressionSet}}
}
