% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser.R
\name{parseProgenesisFeatureCsv}
\alias{parseProgenesisFeatureCsv}
\title{Parse Progenesis Feature Csv Export}
\usage{
parseProgenesisFeatureCsv(
  file = file,
  expDesign = getExpDesignProgenesisCsv(file),
  method = "auc"
)
}
\arguments{
\item{file}{path to Progenesis Feature csv file}

\item{expDesign}{experimental design data.frame}

\item{method}{auc (area under curve) or spc (spectral count)}
}
\value{
ExpressionSet object
}
\description{
Parse Progenesis Feature Csv Export
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
\code{\link[Biobase]{ExpressionSet}}
}
