% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser.R
\name{parseScaffoldRawFile}
\alias{parseScaffoldRawFile}
\title{Parse scaffold output .xls file (RAW export)}
\usage{
parseScaffoldRawFile(
  file,
  expDesign = expDesign,
  keepFirstAcOnly = FALSE,
  isPurityCorrect = T
)
}
\arguments{
\item{file}{path to Scaffold file}

\item{expDesign}{experimental design data.frame}

\item{keepFirstAcOnly}{TRUE/FALSE If multiple ACs in Accession.Numbers filed. Then keep the first one only}

\item{isPurityCorrect}{do purity correction}
}
\value{
ExpressionSet object
}
\description{
Parse scaffold output .xls file (RAW export)
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
\code{\link[Biobase]{ExpressionSet}}
}
