% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser.R
\name{parseSpectronautProteinGroupTxt}
\alias{parseSpectronautProteinGroupTxt}
\title{Parse Spectronaut Protein Group Txt}
\usage{
parseSpectronautProteinGroupTxt(
  file = file,
  expDesign = expDesign,
  method = "auc"
)
}
\arguments{
\item{file}{path to Spectronaut Protein txt file}

\item{expDesign}{experimental design data.frame}

\item{method}{auc (area under curve) or spc (spectral count)}
}
\value{
ExpressionSet object
}
\description{
Parse Spectronaut Protein Group Txt
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")

}
\references{
NA
}
\seealso{
\code{\link[Biobase]{ExpressionSet}}
}
