% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGGraphics.R
\name{plot_violin}
\alias{plot_violin}
\title{Plots violin for all the conditions}
\usage{
plot_violin(eset, fType, ...)
}
\arguments{
\item{eset}{SafeQuantAnalysis object}

\item{fType}{defines from which software the data originally were analyzed}

\item{see}{ggVolcanoPlot}
}
\value{
ggplot2 object
}
\description{
Plots violin for all the conditions
}
\details{
data.frame input object should contain columns ("ratio","pValue","geneName","ac","cv", "description")
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
