% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotLogo}
\alias{plotLogo}
\title{Plot sequence logo}
\usage{
plotLogo(
  motif,
  bgPeptides = "ACDEFGHIKLMNPQRSTVWY",
  main = "",
  targetResidues = c("S", "T", "Y"),
  ic.scale = F,
  ...
)
}
\arguments{
\item{motif}{list of target residue centered motfis}

\item{bgPeptides}{peptides used to calculate residue background frequency (default uniform )}

\item{main}{see plot}

\item{targetResidues}{default \link{STY}}

\item{ic.scale}{logical. If TRUE, the height of each column is proportional to its information content. Otherwise, all columns have the same height.}
}
\description{
Plot sequence logo
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
