% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotMSSignalDistributions}
\alias{plotMSSignalDistributions}
\title{Plot ms.signal distributions}
\usage{
plotMSSignalDistributions(
  d,
  col = 1:100,
  ylab = "Frequnecy",
  xlab = "MS-Signal",
  ...
)
}
\arguments{
\item{d}{matrix of ms-signals}

\item{col}{color}

\item{ylab}{default "Frequnecy"}

\item{xlab}{default "MS-Signal"}

\item{...}{see plot}
}
\description{
Plot ms.signal distributions
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
