% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotNbValidDeFeaturesPerFDR}
\alias{plotNbValidDeFeaturesPerFDR}
\title{Plot Total Number of diffrentially Abundant Features (applying ratio cutoff) vs. qValue/pValue for all conditions}
\usage{
plotNbValidDeFeaturesPerFDR(
  sqa,
  upRegulated = T,
  log2RatioCufOff = log2(1),
  pvalCutOff = 1,
  isLegend = T,
  isAdjusted = T,
  ylab = "Nb. Features",
  xlim = NA,
  ylim = NA,
  ...
)
}
\arguments{
\item{sqa}{SafeQuantAnalysis Object}

\item{upRegulated}{TRUE/FALSE select for upregulated features}

\item{log2RatioCufOff}{log2 ratio cut-off}

\item{pvalCutOff}{pValue/qValue cut-off}

\item{isLegend}{TRUE/FALSE display legend}

\item{isAdjusted}{TRUE/FALSE qValues/pValue on x-axis}

\item{ylab}{default  Nb. Features}

\item{xlim}{see plot}

\item{ylim}{see plot}

\item{...}{see plot}
}
\description{
Plot Total Number of diffrentially Abundant Features (applying ratio cutoff) vs. qValue/pValue for all conditions
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
