% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotROC}
\alias{plotROC}
\title{Plot Number of Identifications vs. FDR}
\usage{
plotROC(
  qvals,
  qvalueThrs = 0.01,
  xlab = "False Discovery Rate",
  ylab = "Nb. Valid Identifications",
  xlim = c(0, 0.1),
  breaks = 100,
  col = "blue",
  lwd = 1.5,
  ...
)
}
\arguments{
\item{qvals}{vector of q-values}

\item{qvalueThrs}{threshold indicated by vertical line}

\item{xlab}{default "False Discovery Rate"}

\item{ylab}{default "Nb. Valid Identifications"}

\item{xlim}{default c(0,0.1)}

\item{breaks}{see breaks for hist function}

\item{col}{default blue}

\item{lwd}{default 1.5}

\item{...}{see plot}
}
\description{
Plot Number of Identifications vs. FDR
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
