% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotScoreDistrib}
\alias{plotScoreDistrib}
\title{Plot identifications target decoy distribution}
\usage{
plotScoreDistrib(
  targetScores,
  decoyScores,
  xlab = "Identification Score",
  ylab = "Counts",
  ...
)
}
\arguments{
\item{targetScores}{target Scores}

\item{decoyScores}{decoy Scores}

\item{xlab}{default "Identification Score"}

\item{ylab}{default "Counts"}

\item{...}{see plot}
}
\description{
Plot identifications target decoy distribution
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
