% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGGraphics.R
\name{plotSelectedProteins2}
\alias{plotSelectedProteins2}
\title{Plots the trends of selected proteins from all condition}
\usage{
plotSelectedProteins2(sqa, isAdjusted = T, ...)
}
\arguments{
\item{sqa}{SafeQuantAnalysis object}

\item{isAdjusted}{(T/F) plot adjusted pvalues}

\item{see}{ggVolcanoPlot}
}
\value{
ggplot2 object
}
\description{
Plots the trends of selected proteins from all condition
}
\details{
data.frame input object should contain columns ("ratio","pValue","geneName","ac","cv", "description")
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
