% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotXYDensity}
\alias{plotXYDensity}
\title{Scatter plot with density coloring}
\usage{
plotXYDensity(
  x,
  y,
  isFitLm = T,
  legendPos = "bottomright",
  disp = c("abline", "R", "Rc"),
  pch = 20,
  ...
)
}
\arguments{
\item{x}{number vector}

\item{y}{number vector}

\item{isFitLm}{fit linear model}

\item{legendPos}{see legend}

\item{disp}{c("abline","R","Rc") display options}

\item{pch}{see plot}

\item{...}{see plot}
}
\description{
Scatter plot with density coloring
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
