% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{stripACs}
\alias{stripACs}
\title{strip uniprot format e.g. "sp|Q8CHJ2|AQP12_MOUSE" ->  Q8CHJ2}
\usage{
stripACs(acs)
}
\arguments{
\item{acs}{accession numbers}
}
\value{
vector character
}
\description{
strip uniprot format e.g. "sp|Q8CHJ2|AQP12_MOUSE" ->  Q8CHJ2
}
\details{
TRUE if less than 10\% of ACs contain a "|" character
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
